+++ app/models/pharmacy.rb
@@ -12,4 +12,5 @@ class Pharmacy < VoidableRecord
   belongs_to :dispensation, class_name: 'Observation',
                             foreign_key: :dispensation_obs_id,
                             optional: true
+  belongs_to :user, foreign_key: :creator
 end